♻️ Move firebase util to root 🔧 [oembed] Fix babel config
